VTN-6758 move formatTime to helper file
