+++ packages/veritone-react-common/src/components/Parts/Buttons/PillButton/styles.scss
@@ -37,6 +37,6 @@
 }
 
 .highlight {
-	border-color: $blue-1;
-	background-color: #DBF0FE;
+  border-color: $blue-1;
+  background-color: #dbf0fe;
 }
