+++ packages/veritone-react-common/src/helpers/time.js
@@ -1,14 +1,23 @@
-export function msToReadableString (milliseconds) {
-  let h, m, s;
-  s = Math.floor(milliseconds / 1000);
-  m = Math.floor(s / 60);
-  s = s % 60;
-  h = Math.floor(m / 60);
-  m = m % 60;
+export function msToReadableString (milliseconds, showDays=false, showMillisec = false) {
+  let dateObject = new Date(milliseconds);
+  let numDays = dateObject.getUTCDate() - 1;
+  let numHours = dateObject.getUTCHours() + (!showDays ? numDays * 24 : 0);
+  let numMinites = dateObject.getUTCMinutes();
+  let numSeconds = dateObject.getUTCSeconds();
+  
 
-  h = (h < 10) && (h > 0) ? "0" + h : h;
-  m = (m < 10) ? "0" + m : m;
-  s = (s < 10) ? "0" + s : s;
+  let showHours = showDays || numHours > 0;
 
-  return (h > 0 ? h + ":" : "" ) + m + ":" + s;
+  let formatedString = '';
+  if (showDays)   formatedString = formatedString + numDays.toLocaleString(undefined, {minimumIntegerDigits: 2}) + ':';
+  if (showHours)  formatedString = formatedString + numHours.toLocaleString(undefined, {minimumIntegerDigits: 2}) + ':';
+  formatedString = formatedString + numMinites.toLocaleString(undefined, {minimumIntegerDigits: 2}) + ':';
+  formatedString = formatedString + numSeconds.toLocaleString(undefined, {minimumIntegerDigits: 2});
+
+  if (showMillisec) {
+    let numMilliseconds = dateObject.getUTCMilliseconds();
+    formatedString = formatedString + ':' + numMilliseconds.toLocaleString(undefined, {minimumIntegerDigits: 3});
+  }
+  
+  return formatedString;
 }
